requirements.txt
fastapi==0.116.0
Pillow==11.3.0
torch==2.7.1+rocm6.2.4
torchvision==0.22.1+rocm6.2.4
